+++ leda/components/DropZone/DropZone.tsx
@@ -16,6 +16,7 @@ import {
 import { useCustomElements, useDropZoneRestProps } from './hooks';
 import { useValidation } from '../Validation';
 import { EMPTY_DROP_ZONE_FILES } from './constants';
+import { LoaderComponent } from '../../src/LoaderComponent';
 
 export const DropZone = React.forwardRef((props: DropZoneProps, ref: React.Ref<DropZoneRefCurrent>): React.ReactElement => {
   const {
@@ -24,7 +25,10 @@ export const DropZone = React.forwardRef((props: DropZoneProps, ref: React.Ref<D
     dropZoneFilesNode,
     forbiddenFiles,
     isDisabled,
+    isLoading,
     isRequired,
+    loadingProgress,
+    loadingViewRender,
     maxFileSize = MAX_FILE_SIZE,
     maxFilesNumber,
     minFileSize = MIN_FILE_SIZE,
@@ -100,6 +104,7 @@ export const DropZone = React.forwardRef((props: DropZoneProps, ref: React.Ref<D
       >
         <Div
           {...rootProps}
+          shouldRender={!isLoading}
           onClick={handleClick}
           className={combinedContentClassNames}
           ref={(component) => {
@@ -135,6 +140,11 @@ export const DropZone = React.forwardRef((props: DropZoneProps, ref: React.Ref<D
             </DescriptionMessage>
           </Info>
         </Div>
+        <LoaderComponent
+          isLoading={isLoading}
+          loadingProgress={loadingProgress}
+          loadingViewRender={loadingViewRender}
+        />
       </Wrapper>
       <InvalidMessage />
       <RejectedFiles